bluetooth/characteristic/writeValue/buffer-is-detached.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';

function detachBuffer(buffer) {
  window.postMessage('', '*', [buffer]);
}

bluetooth_test(async () => {
  let {characteristic, fake_characteristic} =
      await getMeasurementIntervalCharacteristic();

  let typed_array = Uint8Array.of(1, 2);
  detachBuffer(typed_array.buffer);
  try {
    await characteristic.writeValue(typed_array);
    assert_unreached();
  } catch (e) {
    assert_equals(e.code, DOMException.INVALID_STATE_ERR, e.toString());
  }

  let array_buffer = Uint8Array.of(3, 4).buffer;
  detachBuffer(array_buffer);
  try {
    await characteristic.writeValue(array_buffer);
    assert_unreached();
  } catch (e) {
    assert_equals(e.code, DOMException.INVALID_STATE_ERR, e.toString());
  }
}, 'writeValue() fails when passed a detached buffer');
</script>
